Added basic UI and App deployment
Added a basic GUI for picking the file path, and progress bars. Removed plotting of the time alignment of microphones and speakers (but it's just commented). Attempted compiling of fd_conv to C, to speed up processing, but it's currently giving complex output, so I need to actually look at what it's doing. The progressbar UI is still somewhat buggy.